Moved consts to constants.  Import them using a relative path instead. That way we avoid the type error
