+++ src/components/BarChart/components/Bar/Bar.tsx
@@ -1,10 +1,10 @@
-import {MIN_BAR_HEIGHT} from 'consts';
 import React from 'react';
 import {animated, useSpring} from 'react-spring';
 import tokens from '@shopify/polaris-tokens';
 import {Color} from 'types';
 import {ScaleLinear} from 'd3-scale';
 
+import {MIN_BAR_HEIGHT} from '../../../../constants';
 import {getColorValue} from '../../../../utilities';
 
 interface Props {
