+++ src/components/BarChart/contants.ts
@@ -1,0 +1,1 @@
+export const MIN_Y_LABEL_SPACE = 80;
