+++ src/components/GroupedBarChart/components/BarGroup/tests/BarGroup.test.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import {mount} from '@shopify/react-testing';
 import {scaleLinear} from 'd3-scale';
 import {Color} from 'types';
-import {MIN_BAR_HEIGHT} from 'consts';
 
 import {BAR_SPACING} from '../../../constants';
+import {MIN_BAR_HEIGHT} from '../../../../../constants';
 import {BarGroup} from '../BarGroup';
 
 jest.mock('d3-scale', () => ({
