+++ src/components/GroupedBarChart/constants.ts
@@ -2,3 +2,4 @@ export const FULL_OPACITY = '100%';
 export const SUBDUED_OPACITY = '75%';
 export const DEFAULT_HEIGHT = 250;
 export const BAR_SPACING = 0.5;
+export const MIN_Y_LABEL_SPACE = 80;
