+++ src/components/LineChart/hooks/use-y-scale.ts
@@ -1,7 +1,7 @@
-import {SPACING_LOOSE} from 'consts';
 import {useMemo, useEffect, useState} from 'react';
 import {scaleLinear} from 'd3-scale';
 
+import {SPACING_LOOSE} from '../../../constants';
 import {yAxisMinMax} from '../utilities';
 import {MIN_Y_LABEL_SPACE} from '../constants';
 import {Series} from '../types';
