+++ src/constants.ts
@@ -5,3 +5,11 @@ export const SMALL_WIDTH = 300;
 export const MIN_LABEL_SPACE = 100;
 export const TICK_SIZE = 6;
 export const FONT_SIZE = 12;
+export const MIN_BAR_HEIGHT = 5;
+export const MARGIN = {Top: 5, Right: 20, Bottom: 25, Left: 0};
+export const LINE_HEIGHT = 12;
+export const SPACING_EXTRA_TIGHT = 4;
+export const SPACING = 16;
+export const SPACING_LOOSE = 20;
+export const SMALL_FONT_SIZE = 10;
+export const DIAGONAL_ANGLE = -40;
