--- src/consts.ts
@@ -1,13 +1,0 @@
-export const MIN_BAR_HEIGHT = 5;
-export const MARGIN = {Top: 5, Right: 20, Bottom: 25, Left: 0};
-export const LINE_HEIGHT = 12;
-export const TICK_SIZE = 6;
-export const MIN_Y_LABEL_SPACE = 80;
-export const SPACING_EXTRA_TIGHT = 4;
-export const SPACING_TIGHT = 8;
-export const SPACING = 16;
-export const SPACING_LOOSE = 20;
-export const FONT_SIZE = 12;
-export const SMALL_FONT_SIZE = 10;
-export const DIAGONAL_ANGLE = -40;
-export const SMALL_WIDTH = 500;
